remove buffer size
