+++ exporter/sentryexporter/factory.go
@@ -43,8 +43,7 @@ func (f *Factory) CreateDefaultConfig() configmodels.Exporter {
 			TypeVal: typeStr,
 			NameVal: typeStr,
 		},
-		DSN:        "",
-		BufferSize: 30,
+		DSN: "",
 	}
 }
